@@ -1,4 +1,4 @@
-// Code generated by mockery v2.12.1. DO NOT EDIT.
+// Code generated by mockery v2.12.0. DO NOT EDIT.
 
 package annotations
 
@@ -79,6 +79,20 @@ func (_m *FakeAnnotationsRepo) Save(ctx context.Context, item *Item) error {
 	var r0 error
 	if rf, ok := ret.Get(0).(func(context.Context, *Item) error); ok {
 		r0 = rf(ctx, item)
+	} else {
+		r0 = ret.Error(0)
+	}
+
+	return r0
+}
+
+// SaveMany provides a mock function with given fields: ctx, items
+func (_m *FakeAnnotationsRepo) SaveMany(ctx context.Context, items []Item) error {
+	ret := _m.Called(ctx, items)
+
+	var r0 error
+	if rf, ok := ret.Get(0).(func(context.Context, []Item) error); ok {
+		r0 = rf(ctx, items)
 	} else {
 		r0 = ret.Error(0)
 	}
