@@ -19,9 +19,8 @@ import (
 type getStore func(db.DB) store
 
 func testIntegrationGetDashboardVersion(t *testing.T, fn getStore) {
-	if testing.Short() {
-		t.Skip("skipping integration test")
-	}
+	t.Helper()
+
 	ss := db.InitTestDB(t)
 	dashVerStore := fn(ss)
 
