@@ -13,9 +13,7 @@ import (
 type getStore func(db.DB) store
 
 func testIntegrationPlaylistDataAccess(t *testing.T, fn getStore) {
-	if testing.Short() {
-		t.Skip("skipping integration test")
-	}
+	t.Helper()
 
 	ss := db.InitTestDB(t)
 	playlistStore := fn(ss)
