@@ -8,6 +8,10 @@ import (
 	"strings"
 	"testing"
 
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/mock"
+	"github.com/stretchr/testify/require"
+
 	"github.com/grafana/grafana/pkg/services/dashboards"
 	"github.com/grafana/grafana/pkg/services/featuremgmt"
 	"github.com/grafana/grafana/pkg/services/org"
@@ -15,9 +19,7 @@ import (
 	. "github.com/grafana/grafana/pkg/services/publicdashboards/models"
 	"github.com/grafana/grafana/pkg/services/user"
 	"github.com/grafana/grafana/pkg/setting"
-	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/mock"
-	"github.com/stretchr/testify/require"
+	"github.com/grafana/grafana/pkg/util/errutil"
 )
 
 var userAdmin = &user.SignedInUser{UserID: 1, OrgID: 1, OrgRole: org.RoleAdmin, Login: "testAdminUser"}
@@ -121,7 +123,7 @@ func TestAPIListPublicDashboard(t *testing.T) {
 			Name:                 "Handles Service error",
 			User:                 userViewer,
 			Response:             nil,
-			ResponseErr:          errors.New("error, service broken"),
+			ResponseErr:          ErrInternalServerError.Errorf(""),
 			ExpectedHttpResponse: http.StatusInternalServerError,
 		},
 	}
@@ -148,11 +150,145 @@ func TestAPIListPublicDashboard(t *testing.T) {
 			}
 
 			if test.ResponseErr != nil {
-				var errResp JsonErrResponse
+				var errResp errutil.PublicError
 				err := json.Unmarshal(response.Body.Bytes(), &errResp)
 				require.NoError(t, err)
-				assert.Equal(t, "error, service broken", errResp.Error)
+				assert.Equal(t, "Internal server error", errResp.Message)
+				assert.Equal(t, "publicdashboards.internalServerError", errResp.MessageID)
 				service.AssertNotCalled(t, "FindAll")
+			}
+		})
+	}
+}
+
+func TestAPIDeletePublicDashboard(t *testing.T) {
+	dashboardUid := "abc1234"
+	publicDashboardUid := "1234asdfasdf"
+	userEditorAllPublicDashboard := &user.SignedInUser{UserID: 4, OrgID: 1, OrgRole: org.RoleEditor, Login: "testEditorUser", Permissions: map[int64]map[string][]string{1: {dashboards.ActionDashboardsPublicWrite: {dashboards.ScopeDashboardsAll}}}}
+	userEditorAnotherPublicDashboard := &user.SignedInUser{UserID: 4, OrgID: 1, OrgRole: org.RoleEditor, Login: "testEditorUser", Permissions: map[int64]map[string][]string{1: {dashboards.ActionDashboardsPublicWrite: {"another-uid"}}}}
+	userEditorPublicDashboard := &user.SignedInUser{UserID: 4, OrgID: 1, OrgRole: org.RoleEditor, Login: "testEditorUser", Permissions: map[int64]map[string][]string{1: {dashboards.ActionDashboardsPublicWrite: {fmt.Sprintf("dashboards:uid:%s", dashboardUid)}}}}
+
+	testCases := []struct {
+		Name                    string
+		User                    *user.SignedInUser
+		DashboardUid            string
+		PublicDashboardUid      string
+		ResponseErr             error
+		ExpectedHttpResponse    int
+		ExpectedMessageResponse string
+		ShouldCallService       bool
+	}{
+		{
+			Name:                 "User viewer cannot delete public dashboard",
+			User:                 userViewer,
+			DashboardUid:         dashboardUid,
+			PublicDashboardUid:   publicDashboardUid,
+			ResponseErr:          nil,
+			ExpectedHttpResponse: http.StatusForbidden,
+			ShouldCallService:    false,
+		},
+		{
+			Name:                 "User editor without specific dashboard access cannot delete public dashboard",
+			User:                 userEditorAnotherPublicDashboard,
+			DashboardUid:         dashboardUid,
+			PublicDashboardUid:   publicDashboardUid,
+			ResponseErr:          nil,
+			ExpectedHttpResponse: http.StatusForbidden,
+			ShouldCallService:    false,
+		},
+		{
+			Name:                 "User editor with all dashboard accesses can delete public dashboard",
+			User:                 userEditorAllPublicDashboard,
+			DashboardUid:         dashboardUid,
+			PublicDashboardUid:   publicDashboardUid,
+			ResponseErr:          nil,
+			ExpectedHttpResponse: http.StatusOK,
+			ShouldCallService:    true,
+		},
+		{
+			Name:                 "User editor with dashboard access can delete public dashboard",
+			User:                 userEditorPublicDashboard,
+			DashboardUid:         dashboardUid,
+			PublicDashboardUid:   publicDashboardUid,
+			ResponseErr:          nil,
+			ExpectedHttpResponse: http.StatusOK,
+			ShouldCallService:    true,
+		},
+		{
+			Name:                    "Internal server error returns an error",
+			User:                    userEditorPublicDashboard,
+			DashboardUid:            dashboardUid,
+			PublicDashboardUid:      publicDashboardUid,
+			ResponseErr:             ErrInternalServerError.Errorf(""),
+			ExpectedHttpResponse:    ErrInternalServerError.Errorf("").Reason.Status().HTTPStatus(),
+			ExpectedMessageResponse: ErrInternalServerError.Errorf("").PublicMessage,
+			ShouldCallService:       true,
+		},
+		{
+			Name:                    "PublicDashboard error returns correct status code instead of 500",
+			User:                    userEditorPublicDashboard,
+			DashboardUid:            dashboardUid,
+			PublicDashboardUid:      publicDashboardUid,
+			ResponseErr:             ErrPublicDashboardIdentifierNotSet.Errorf(""),
+			ExpectedHttpResponse:    ErrPublicDashboardIdentifierNotSet.Errorf("").Reason.Status().HTTPStatus(),
+			ExpectedMessageResponse: ErrPublicDashboardIdentifierNotSet.Errorf("").PublicMessage,
+			ShouldCallService:       true,
+		},
+		{
+			Name:                 "Invalid publicDashboardUid throws an error",
+			User:                 userEditorPublicDashboard,
+			DashboardUid:         dashboardUid,
+			PublicDashboardUid:   "inv@lid-publicd@shboard-uid!",
+			ResponseErr:          nil,
+			ExpectedHttpResponse: http.StatusBadRequest,
+			ShouldCallService:    false,
+		},
+		{
+			Name:                    "Public dashboard uid does not exist",
+			User:                    userEditorPublicDashboard,
+			DashboardUid:            dashboardUid,
+			PublicDashboardUid:      "UIDDOESNOTEXIST",
+			ResponseErr:             ErrPublicDashboardNotFound.Errorf(""),
+			ExpectedHttpResponse:    ErrPublicDashboardNotFound.Errorf("").Reason.Status().HTTPStatus(),
+			ExpectedMessageResponse: ErrPublicDashboardNotFound.Errorf("").PublicMessage,
+			ShouldCallService:       true,
+		},
+	}
+
+	for _, test := range testCases {
+		t.Run(test.Name, func(t *testing.T) {
+			service := publicdashboards.NewFakePublicDashboardService(t)
+
+			if test.ShouldCallService {
+				service.On("Delete", mock.Anything, mock.Anything, mock.Anything).
+					Return(test.ResponseErr)
+			}
+
+			cfg := setting.NewCfg()
+
+			features := featuremgmt.WithFeatures(featuremgmt.FlagPublicDashboards)
+			testServer := setupTestServer(t, cfg, features, service, nil, test.User)
+
+			response := callAPI(testServer, http.MethodDelete, fmt.Sprintf("/api/dashboards/uid/%s/public-dashboards/%s", test.DashboardUid, test.PublicDashboardUid), nil, t)
+			assert.Equal(t, test.ExpectedHttpResponse, response.Code)
+
+			if test.ExpectedHttpResponse == http.StatusOK {
+				var jsonResp any
+				err := json.Unmarshal(response.Body.Bytes(), &jsonResp)
+				require.NoError(t, err)
+				assert.Equal(t, jsonResp, nil)
+			}
+
+			if !test.ShouldCallService {
+				service.AssertNotCalled(t, "Delete")
+			}
+
+			if test.ResponseErr != nil {
+				var errResp errutil.PublicError
+				err := json.Unmarshal(response.Body.Bytes(), &errResp)
+				require.NoError(t, err)
+				assert.Equal(t, test.ExpectedHttpResponse, errResp.StatusCode)
+				assert.Equal(t, test.ExpectedMessageResponse, errResp.Message)
 			}
 		})
 	}
@@ -186,7 +322,7 @@ func TestAPIGetPublicDashboard(t *testing.T) {
 			DashboardUid:          "77777",
 			ExpectedHttpResponse:  http.StatusNotFound,
 			PublicDashboardResult: nil,
-			PublicDashboardErr:    dashboards.ErrDashboardNotFound,
+			PublicDashboardErr:    ErrDashboardNotFound.Errorf(""),
 			User:                  userViewer,
 			AccessControlEnabled:  false,
 			ShouldCallService:     true,
@@ -288,7 +424,7 @@ func TestApiCreatePublicDashboard(t *testing.T) {
 			Name:                 "returns 500 when not persisted",
 			ExpectedHttpResponse: http.StatusInternalServerError,
 			publicDashboard:      &PublicDashboard{},
-			SaveDashboardErr:     errors.New("backend failed to save"),
+			SaveDashboardErr:     ErrInternalServerError.Errorf(""),
 			User:                 userAdmin,
 			AccessControlEnabled: false,
 			ShouldCallService:    true,
@@ -297,7 +433,7 @@ func TestApiCreatePublicDashboard(t *testing.T) {
 			Name:                 "returns 404 when dashboard not found",
 			ExpectedHttpResponse: http.StatusNotFound,
 			publicDashboard:      &PublicDashboard{},
-			SaveDashboardErr:     dashboards.ErrDashboardNotFound,
+			SaveDashboardErr:     ErrDashboardNotFound.Errorf(""),
 			User:                 userAdmin,
 			AccessControlEnabled: false,
 			ShouldCallService:    true,
@@ -316,7 +452,7 @@ func TestApiCreatePublicDashboard(t *testing.T) {
 			Name:                 "returns 403 when no permissions",
 			ExpectedHttpResponse: http.StatusForbidden,
 			publicDashboard:      &PublicDashboard{IsEnabled: true},
-			SaveDashboardErr:     nil,
+			SaveDashboardErr:     ErrInternalServerError.Errorf("default error"),
 			User:                 userViewer,
 			AccessControlEnabled: false,
 			ShouldCallService:    false,
@@ -400,7 +536,7 @@ func TestAPIUpdatePublicDashboard(t *testing.T) {
 			DashboardUid:         "",
 			PublicDashboardUid:   "",
 			PublicDashboardRes:   nil,
-			PublicDashboardErr:   dashboards.ErrDashboardIdentifierInvalid,
+			PublicDashboardErr:   ErrPublicDashboardIdentifierNotSet.Errorf(""),
 			ExpectedHttpResponse: http.StatusNotFound,
 			ShouldCallService:    false,
 		},
@@ -410,7 +546,7 @@ func TestAPIUpdatePublicDashboard(t *testing.T) {
 			DashboardUid:         dashboardUid,
 			PublicDashboardUid:   "",
 			PublicDashboardRes:   nil,
-			PublicDashboardErr:   ErrPublicDashboardNotFound,
+			PublicDashboardErr:   ErrPublicDashboardNotFound.Errorf(""),
 			ExpectedHttpResponse: http.StatusNotFound,
 			ShouldCallService:    false,
 		},
@@ -420,7 +556,7 @@ func TestAPIUpdatePublicDashboard(t *testing.T) {
 			DashboardUid:         dashboardUid,
 			PublicDashboardUid:   publicDashboardUid,
 			PublicDashboardRes:   nil,
-			PublicDashboardErr:   dashboards.ErrDashboardNotFound,
+			PublicDashboardErr:   ErrDashboardNotFound.Errorf(""),
 			ExpectedHttpResponse: http.StatusNotFound,
 			ShouldCallService:    true,
 		},
@@ -501,134 +637,6 @@ func TestAPIUpdatePublicDashboard(t *testing.T) {
 				err := json.Unmarshal(response.Body.Bytes(), &errResp)
 				require.NoError(t, err)
 				assert.Equal(t, test.PublicDashboardErr.Error(), errResp.Error)
-			}
-		})
-	}
-}
-
-func TestAPIDeletePublicDashboard(t *testing.T) {
-	dashboardUid := "abc1234"
-	publicDashboardUid := "1234asdfasdf"
-	userEditorAllPublicDashboard := &user.SignedInUser{UserID: 4, OrgID: 1, OrgRole: org.RoleEditor, Login: "testEditorUser", Permissions: map[int64]map[string][]string{1: {dashboards.ActionDashboardsPublicWrite: {dashboards.ScopeDashboardsAll}}}}
-	userEditorAnotherPublicDashboard := &user.SignedInUser{UserID: 4, OrgID: 1, OrgRole: org.RoleEditor, Login: "testEditorUser", Permissions: map[int64]map[string][]string{1: {dashboards.ActionDashboardsPublicWrite: {"another-uid"}}}}
-	userEditorPublicDashboard := &user.SignedInUser{UserID: 4, OrgID: 1, OrgRole: org.RoleEditor, Login: "testEditorUser", Permissions: map[int64]map[string][]string{1: {dashboards.ActionDashboardsPublicWrite: {fmt.Sprintf("dashboards:uid:%s", dashboardUid)}}}}
-
-	testCases := []struct {
-		Name                 string
-		User                 *user.SignedInUser
-		DashboardUid         string
-		PublicDashboardUid   string
-		ResponseErr          error
-		ExpectedHttpResponse int
-		ShouldCallService    bool
-	}{
-		{
-			Name:                 "User viewer cannot delete public dashboard",
-			User:                 userViewer,
-			DashboardUid:         dashboardUid,
-			PublicDashboardUid:   publicDashboardUid,
-			ResponseErr:          nil,
-			ExpectedHttpResponse: http.StatusForbidden,
-			ShouldCallService:    false,
-		},
-		{
-			Name:                 "User editor without specific dashboard access cannot delete public dashboard",
-			User:                 userEditorAnotherPublicDashboard,
-			DashboardUid:         dashboardUid,
-			PublicDashboardUid:   publicDashboardUid,
-			ResponseErr:          nil,
-			ExpectedHttpResponse: http.StatusForbidden,
-			ShouldCallService:    false,
-		},
-		{
-			Name:                 "User editor with all dashboard accesses can delete public dashboard",
-			User:                 userEditorAllPublicDashboard,
-			DashboardUid:         dashboardUid,
-			PublicDashboardUid:   publicDashboardUid,
-			ResponseErr:          nil,
-			ExpectedHttpResponse: http.StatusOK,
-			ShouldCallService:    true,
-		},
-		{
-			Name:                 "User editor with dashboard access can delete public dashboard",
-			User:                 userEditorPublicDashboard,
-			DashboardUid:         dashboardUid,
-			PublicDashboardUid:   publicDashboardUid,
-			ResponseErr:          nil,
-			ExpectedHttpResponse: http.StatusOK,
-			ShouldCallService:    true,
-		},
-		{
-			Name:                 "Internal server error returns an error",
-			User:                 userEditorPublicDashboard,
-			DashboardUid:         dashboardUid,
-			PublicDashboardUid:   publicDashboardUid,
-			ResponseErr:          errors.New("server error"),
-			ExpectedHttpResponse: http.StatusInternalServerError,
-			ShouldCallService:    true,
-		},
-		{
-			Name:                 "PublicDashboard error returns correct status code instead of 500",
-			User:                 userEditorPublicDashboard,
-			DashboardUid:         dashboardUid,
-			PublicDashboardUid:   publicDashboardUid,
-			ResponseErr:          ErrPublicDashboardIdentifierNotSet,
-			ExpectedHttpResponse: ErrPublicDashboardIdentifierNotSet.StatusCode,
-			ShouldCallService:    true,
-		},
-		{
-			Name:                 "Invalid publicDashboardUid throws an error",
-			User:                 userEditorPublicDashboard,
-			DashboardUid:         dashboardUid,
-			PublicDashboardUid:   "inv@lid-publicd@shboard-uid!",
-			ResponseErr:          nil,
-			ExpectedHttpResponse: ErrPublicDashboardIdentifierNotSet.StatusCode,
-			ShouldCallService:    false,
-		},
-		{
-			Name:                 "Public dashboard uid does not exist",
-			User:                 userEditorPublicDashboard,
-			DashboardUid:         dashboardUid,
-			PublicDashboardUid:   "UIDDOESNOTEXIST",
-			ResponseErr:          ErrPublicDashboardNotFound,
-			ExpectedHttpResponse: ErrPublicDashboardNotFound.StatusCode,
-			ShouldCallService:    true,
-		},
-	}
-
-	for _, test := range testCases {
-		t.Run(test.Name, func(t *testing.T) {
-			service := publicdashboards.NewFakePublicDashboardService(t)
-
-			if test.ShouldCallService {
-				service.On("Delete", mock.Anything, mock.Anything, mock.Anything).
-					Return(test.ResponseErr)
-			}
-
-			cfg := setting.NewCfg()
-
-			features := featuremgmt.WithFeatures(featuremgmt.FlagPublicDashboards)
-			testServer := setupTestServer(t, cfg, features, service, nil, test.User)
-
-			response := callAPI(testServer, http.MethodDelete, fmt.Sprintf("/api/dashboards/uid/%s/public-dashboards/%s", test.DashboardUid, test.PublicDashboardUid), nil, t)
-			assert.Equal(t, test.ExpectedHttpResponse, response.Code)
-
-			if test.ExpectedHttpResponse == http.StatusOK {
-				var jsonResp any
-				err := json.Unmarshal(response.Body.Bytes(), &jsonResp)
-				require.NoError(t, err)
-				assert.Equal(t, jsonResp, nil)
-			}
-
-			if !test.ShouldCallService {
-				service.AssertNotCalled(t, "Delete")
-			}
-
-			if test.ResponseErr != nil {
-				var errResp JsonErrResponse
-				err := json.Unmarshal(response.Body.Bytes(), &errResp)
-				require.NoError(t, err)
-				assert.Equal(t, test.ResponseErr.Error(), errResp.Error)
 			}
 		})
 	}
