@@ -12,7 +12,7 @@ import (
 	"github.com/grafana/grafana/pkg/tsdb/grafanads"
 )
 
-func TestCreateAndDelete(t *testing.T) {
+func TestIntegrationCreateAndDelete(t *testing.T) {
 	if testing.Short() {
 		t.Skip("skipping integration test")
 	}
@@ -123,7 +123,7 @@ func createQuery(t *testing.T, ctx context.Context, testCtx testContext) string 
 	return search[0].uid
 }
 
-func TestDashboardGetWithLatestSavedQueries(t *testing.T) {
+func TestIntegrationDashboardGetWithLatestSavedQueries(t *testing.T) {
 	if testing.Short() {
 		t.Skip("skipping integration test")
 	}
